feat(abilities): ground-target abilities apply true AoE damage to all enemies in radius
Closes #81. The previous handle_use_ability_on_ground picked the
nearest hostile NPC within a hardcoded 5.0m radius and applied
single-target damage. For grenades / mortars / AoE drops this was
wrong — every other enemy in the blast got nothing while the player
paid the cooldown.

Implementation:

  cell::abilities::damage_apply (NEW)
    Extracted the per-target damage block from handle_use_ability
    (~280 lines) into a reusable apply_damage_to_target helper:
    snapshot attacker stats → fresh QR roll → apply health/focus
    damage → death side effects (death-transition burst, kill XP,
    onBeginAidWait, onSequence death anim) → threat generation on
    survival. Takes a `needs_ammo_stat_send` flag so callers control
    whether to flush the bandolier ammo stat (primary call: yes;
    AoE secondary: no — the primary already flushed).

  cell::abilities::use_ability
    handle_use_ability now delegates the post-validation damage
    path to damage_apply::apply_damage_to_target. The function
    dropped from 537 → 264 lines. The validation / consume / timer
    / sequence / state-field plumbing stayed inline because it
    fires regardless of whether a target was picked.

  cell::abilities::dispatch
    handle_use_ability_on_ground rewritten:
      - ability_radius() reads the AoE radius from the ability's
        first effect_def's `Radius` NVP (default 5.0 if missing,
        preserving prior behavior for abilities without authored
        AoE data).
      - Collects all hostile NPCs within radius_sq, sorted by
        distance from the click point. Hostile-faction sentinel
        unchanged (faction == 10, !is_player).
      - Primary range-check unchanged: nearest target must also be
        within attacker's ability max_range, else fall through to
        cooldown-only path.
      - Primary target: full handle_use_ability path (consumes
        ammo + cooldown).
      - Secondary targets: apply_damage_to_target with
        needs_ammo_stat_send=false and a fresh effect_seq per
        target (independent QR rolls, independent client packet
        correlation).

Acceptance per #81:
  ✓ Damages every enemy in the configured radius, not just nearest
  ✓ Radius is data-driven from the effect definition (Radius NVP)
  ✓ Cooldown and ammo consumed once per invocation
  ✓ Combat-state transitions (BSF_InCombat, BSF_Holster clear)
    fire once via the primary handle_use_ability path
  ✓ Single-target abilities continue to work — handle_use_ability's
    delegation to apply_damage_to_target is the same call shape

Out of scope (deferred):
  - friendly_fire flag handling. The Ability/Effect schema doesn't
    carry a friendly_fire bit yet; AoE today is hostile-only by
    the faction filter.
  - Combined onEffectResults aggregation (one packet listing all
    targets vs N per-target packets). Current is N — matches python
    reference for single-target hits and seems to render fine.
  - AoE secondary kills don't fire the content-engine entity_death
    chain. The python reference fired onDead per target via the
    cell-side hook in handle_interact's combat fall-through; AoE
    paths skip that fall-through. Worth a follow-up if any chain
    triggers off mass-kill events (none currently shipped).

Tests:
  - ability_radius_reads_effect_nvp — synthesizes an ability + effect
    with a Radius NVP, asserts ability_radius reads it.
  - ability_radius_falls_back_to_default_when_unset — None ability
    → 5.0.
  - ability_radius_falls_back_when_effects_have_no_radius — effect
    has HealthDamage but no Radius → 5.0.

  cargo test -p cimmeria-services --lib abilities: 8 passed (was 5,
  +3 new). Full services suite: 250 passed.
